backend/requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# CORS
django-cors-headers>=4.3,<5.0

# Database
psycopg[binary]>=3.1,<4.0

# Cache & Message Broker
redis>=5.0,<6.0

# Task Queue
celery[redis]>=5.3,<6.0
django-celery-beat>=2.5,<3.0

# Production Server
gunicorn>=21.0,<23.0

# Environment
python-decouple>=3.8,<4.0

# HTTP Client
requests>=2.31,<3.0

# Development & Testing
ruff>=0.1.0
pytest>=7.4,<9.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<6.0
